addMeteoriteList
