+++ app/src/main/java/com/antonio/samir/meteoritelandingsspots/features/list/ListScreen.kt
@@ -1,18 +1,17 @@
+package com.antonio.samir.meteoritelandingsspots.features.list
+
+
 import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -22,25 +21,14 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import androidx.paging.LoadState
 import androidx.paging.PagingData
-import androidx.paging.compose.collectAsLazyPagingItems
-import androidx.paging.compose.items
-import com.antonio.samir.meteoritelandingsspots.R
 import com.antonio.samir.meteoritelandingsspots.common.ResultOf
 import com.antonio.samir.meteoritelandingsspots.designsystem.ui.components.AddressProgress
 import com.antonio.samir.meteoritelandingsspots.designsystem.ui.components.Header
 import com.antonio.samir.meteoritelandingsspots.designsystem.ui.components.HeaderState
-import com.antonio.samir.meteoritelandingsspots.designsystem.ui.components.Loading
-import com.antonio.samir.meteoritelandingsspots.designsystem.ui.components.MessageError
-import com.antonio.samir.meteoritelandingsspots.designsystem.ui.components.MeteoriteCell
 import com.antonio.samir.meteoritelandingsspots.designsystem.ui.components.MeteoriteItemView
 import com.antonio.samir.meteoritelandingsspots.designsystem.ui.theme.MeteoriteLandingsTheme
-import com.antonio.samir.meteoritelandingsspots.features.list.ListScreenView
-import com.antonio.samir.meteoritelandingsspots.features.list.ListState
-import com.antonio.samir.meteoritelandingsspots.features.list.MeteoriteListViewModel
 import kotlinx.coroutines.FlowPreview
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
 
 @OptIn(FlowPreview::class)
@@ -57,7 +45,9 @@ fun ListScreenNavigation(
         uiState = uiState,
         onItemClick = viewModel::selectMeteorite,
         onEnterSearch = { viewModel.setHeaderState(HeaderState.Search) },
-        onExitSearch = {},
+        onExitSearch = {
+            viewModel.setHeaderState(HeaderState.Expanded)
+        },
         onTopList = viewModel::onTopList,
         onSearch = {}
     )
@@ -77,8 +67,6 @@ fun ListScreen(
 ) {
 
     val scrollState = rememberLazyListState()
-
-    val scrollBehavior = TopAppBarDefaults
 
     val scrollOffset: Float = Math.min(
         1f,
@@ -122,10 +110,11 @@ fun ListScreen(
                             meteorites = uiState.listState.meteorites,
                             onItemClick = onItemClick
                         )
-                        ListState.UiLoading -> Loading(modifier = Modifier.fillMaxSize())
-                        is ListState.UiMessage -> MessageError(
-                            message = uiState.listState.message
-                        )
+//                        ListState.UiLoading -> Loading(modifier = Modifier.fillMaxSize())
+//                        is ListState.UiMessage -> MessageError(
+//                            message = uiState.listState.message
+//                        )
+                        else -> {}
                     }
 
                     val addressProgress = uiState.addressStatus
@@ -147,55 +136,6 @@ fun ListScreen(
 }
 
 
-@Composable
-private fun MeteoriteList(
-    scrollState: LazyListState,
-    meteorites: Flow<PagingData<MeteoriteItemView>>,
-    onItemClick: (itemView: MeteoriteItemView) -> Unit
-) {
-    val items = meteorites.collectAsLazyPagingItems()
-    LazyColumn(
-        state = scrollState,
-        modifier = Modifier
-            .fillMaxSize()
-            .background(MaterialTheme.colorScheme.background)
-    ) {
-
-        items(items) { item ->
-            item?.let { MeteoriteCell(it, onItemClick) }
-        }
-
-        items.apply {
-            when {
-                loadState.refresh is LoadState.Loading -> {
-                    item { Loading(modifier = Modifier.fillParentMaxSize()) }
-                }
-                loadState.append is LoadState.Loading -> {
-                    item { Loading(modifier = Modifier.fillParentMaxSize()) }
-                }
-                loadState.refresh is LoadState.Error -> {
-                    val e = items.loadState.refresh as LoadState.Error
-                    item {
-                        MessageError(
-                            message = e.error.localizedMessage!!,
-                            modifier = Modifier.fillParentMaxSize()
-                        )
-                    }
-                }
-                loadState.append is LoadState.Error -> {
-                    val e = items.loadState.append as LoadState.Error
-                    item {
-                        MessageError(
-                            message = e.error.localizedMessage!!
-                        )
-                    }
-                }
-            }
-        }
-
-    }
-}
-
 @ExperimentalAnimationApi
 @ExperimentalComposeUiApi
 @Preview("Meteorite list view")
@@ -214,74 +154,74 @@ fun ListScreenPreview() {
     ) {}
 
 }
-
-@ExperimentalAnimationApi
-@ExperimentalComposeUiApi
-@Preview("Meteorite list loading")
-@Composable
-fun ListScreenLoadingPreview() {
-    ListScreen(
-        uiState = ListScreenView(
-            addressStatus = ResultOf.Success(100f),
-            listState = ListState.UiLoading,
-            onDarkModeToggleClick = { },
-        ),
-        {}
-    ) {}
-
-}
-
-@ExperimentalAnimationApi
-@ExperimentalComposeUiApi
-@Preview("Meteorite list error")
-@Composable
-fun ListScreenErrorPreview() {
-    ListScreen(
-        uiState = ListScreenView(
-            listState = ListState.UiMessage(R.string.general_error),
-            addressStatus = ResultOf.Success(100f),
-            onDarkModeToggleClick = { },
-        ),
-        {}
-    ) {}
-
-}
-
-@ExperimentalAnimationApi
-@ExperimentalComposeUiApi
-@Preview("Meteorite list message loading 10%")
-@Composable
-fun ListScreenMessagePreview() {
-    val items = getFakeListItems()
-
-    ListScreen(
-        uiState = ListScreenView(
-            addressStatus = ResultOf.InProgress(10.0f),
-            listState = ListState.UiContent(meteorites = flowOf(PagingData.from(items))),
-            onDarkModeToggleClick = { }
-        ),
-        onSearch = {}
-    )
-
-}
-
-@Preview
-@Composable
-fun TextLazyColumn() {
-
-    val items = getFakeListItems()
-
-    LazyColumn(Modifier.height(500.dp)) {
-
-        items.forEach {
-            item {
-                MeteoriteCell(it, null)
-            }
-        }
-
-    }
-
-}
+//
+//@ExperimentalAnimationApi
+//@ExperimentalComposeUiApi
+//@Preview("Meteorite list loading")
+//@Composable
+//fun ListScreenLoadingPreview() {
+//    ListScreen(
+//        uiState = ListScreenView(
+//            addressStatus = ResultOf.Success(100f),
+//            listState = ListState.UiLoading,
+//            onDarkModeToggleClick = { },
+//        ),
+//        {}
+//    ) {}
+//
+//}
+//
+//@ExperimentalAnimationApi
+//@ExperimentalComposeUiApi
+//@Preview("Meteorite list error")
+//@Composable
+//fun ListScreenErrorPreview() {
+//    ListScreen(
+//        uiState = ListScreenView(
+//            listState = ListState.UiMessage(R.string.general_error),
+//            addressStatus = ResultOf.Success(100f),
+//            onDarkModeToggleClick = { },
+//        ),
+//        {}
+//    ) {}
+//
+//}
+//
+//@ExperimentalAnimationApi
+//@ExperimentalComposeUiApi
+//@Preview("Meteorite list message loading 10%")
+//@Composable
+//fun ListScreenMessagePreview() {
+//    val items = getFakeListItems()
+//
+//    ListScreen(
+//        uiState = ListScreenView(
+//            addressStatus = ResultOf.InProgress(10.0f),
+//            listState = ListState.UiContent(meteorites = flowOf(PagingData.from(items))),
+//            onDarkModeToggleClick = { }
+//        ),
+//        onSearch = {}
+//    )
+//
+//}
+//
+//@Preview
+//@Composable
+//fun TextLazyColumn() {
+//
+//    val items = getFakeListItems()
+//
+//    LazyColumn(Modifier.height(500.dp)) {
+//
+//        items.forEach {
+//            item {
+//                MeteoriteCell(it, null)
+//            }
+//        }
+//
+//    }
+//
+//}
 
 private fun getFakeListItems(): List<MeteoriteItemView> = (1..10).map {
     MeteoriteItemView(
